refactor code pretaining to saving and loading records
