+++ app/src/main/java/com/example/maths/SpeedMath.kt
@@ -11,18 +11,6 @@ import java.util.Date
 class SpeedMath: Application() {
 
     companion object {
-        var bestTimeHard: Long = 0
-        var bestTimeHardString: String = ""
-        var bestTimeEasy: Long = 0
-        var bestTimeEasyString: String = ""
-        var bestAudioTimeHard: Long = 0
-        var bestAudioTimeEasy: Long = 0
-        var bestAudioTimeHardString: String = ""
-        var bestAudioTimeEasyString: String = ""
-        var currScore = 0
-        var difficulty = 0
-        var gameMode = 0
-        var gameTime: Long = 0
 
         fun formatTime(ms: Long): String {
             val minutes = ms / 1000 /60
@@ -62,36 +50,6 @@ class SpeedMath: Application() {
             AppDatabase::class.java, "sm-database"
         ).allowMainThreadQueries().build()
 
-        try {
-            loadHighScore()
-        } catch (e: FileNotFoundException) {
-            // File not found
-        }
     }
-
-    private fun loadHighScore() {
-        val recordDao = db!!.recordDao()
-        val scores = recordDao.getAllFastestTimes()
-        for (score in scores) {
-            if (score.gameMode == 0) {
-                if (score.difficulty == 0) {
-                    bestTimeEasy = score.gameTime
-                } else {
-                    bestTimeHard = score.gameTime
-                }
-            } else {
-                if (score.difficulty == 0) {
-                    bestAudioTimeEasy = score.gameTime
-                } else{
-                    bestAudioTimeHard = score.gameTime
-                }
-            }
-            bestTimeHardString = formatTime(bestTimeHard)
-            bestTimeEasyString = formatTime(bestTimeEasy)
-            bestAudioTimeEasyString = formatTime(bestAudioTimeEasy)
-            bestAudioTimeHardString = formatTime(bestAudioTimeHard)
-        }
-    }
-
 
 }
